+++ containers/product/product-details.tsx
