+++ containers/search.tsx
@@ -3,7 +3,7 @@ import useSearch from '@framework/product/use-search'
 import { filterQuery, useSearchMeta } from '@lib/search'
 import type { SearchPropsType } from '@lib/search-props'
 import { Grid } from '@mui/material'
-import ProductCardContainer from 'containers/product/ProductCardContainer'
+import ProductCardContainer from 'containers/product/product-card'
 import { useRouter } from 'next/router'
 import * as React from 'react'
 import { useState } from 'react'
